feature(database): migrated main enitites to long id

+++ src/Mapingway.Application/Abstractions/Authentication/IPermissionRepository.cs
@@ -2,5 +2,5 @@
 
 public interface IPermissionRepository
 {
-    Task<HashSet<string>> GetPermissionsAsync(int userId, CancellationToken cancellationToken);
+    Task<HashSet<string>> GetPermissionsAsync(long userId, CancellationToken cancellationToken);
 }
